# docs/openapi-v1-draft.yaml
openapi: 3.0.3
info:
  title: ACI Suite API
  version: 1.0.0-draft
  description: >
    Rascunho OpenAPI v1 baseado na matriz de baseline da Fase 0.
    Backend canônico alvo: Express (`src/backend/server.ts`).
    Endpoints com gap no Express estão marcados com `x-implementation: missing`.
servers:
  - url: http://localhost:4001
    description: Local (Express)
  - url: https://aci.automacoescomerciais.com.br
    description: Produção
...
security:
  - bearerAuth: []
paths:
...
  /api/settings/me:
    get:
      tags: [Settings]
      summary: Obter configurações do usuário autenticado
      responses:
        "200":
          description: Configurações retornadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/settings:
    post:
      tags: [Settings]
      summary: Salvar configurações do usuário autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Configurações salvas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SuccessMessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string